Upgrade to rabbitmq-c v0.5.0

+++ librabbitmq/amqp_timer.c
@@ -22,7 +22,6 @@
  */
 #include "amqp.h"
 #include "amqp_timer.h"
-#include <posix/time.h>
 #include <string.h>
 
 #if (defined(_WIN32) || defined(__WIN32__) || defined(WIN32))
@@ -43,7 +42,7 @@
 uint64_t
 amqp_get_monotonic_timestamp(void)
 {
-  static uint64_t NS_PER_COUNT = 0;
+  static double NS_PER_COUNT = 0;
   LARGE_INTEGER perf_count;
 
   if (0 == NS_PER_COUNT) {
@@ -51,14 +50,14 @@ amqp_get_monotonic_timestamp(void)
     if (!QueryPerformanceFrequency(&perf_frequency)) {
       return 0;
     }
-    NS_PER_COUNT = AMQP_NS_PER_S / perf_frequency.QuadPart;
+    NS_PER_COUNT = (double)AMQP_NS_PER_S / perf_frequency.QuadPart;
   }
 
   if (!QueryPerformanceCounter(&perf_count)) {
     return 0;
   }
 
-  return perf_count.QuadPart * NS_PER_COUNT;
+  return (uint64_t)(perf_count.QuadPart * NS_PER_COUNT);
 }
 #endif /* AMQP_WIN_TIMER_API */
 
@@ -92,13 +91,9 @@ amqp_get_monotonic_timestamp(void)
 uint64_t
 amqp_get_monotonic_timestamp(void)
 {
-  /* The library expects nanoseconds from this function.
-   * AMQP_64BIT_TIME_FUNC() is defined in
-   * lwipdemo/include/config.h  as usecSincePowerOn()
-   */
   return AMQP_NS_PER_US * AMQP_64BIT_TIME_FUNC();
 }
-#endif
+#endif /* AMQP_64BIT_TIME_API */
 
 #ifdef AMQP_POSIX_TIMER_API
 #include <time.h>
